# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Aeko Enhanced API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Aeko Development Team",
      "email": "dev@aeko.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "username": {
            "type": "string",
            "description": "Unique username"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email"
          },
          "profilePicture": {
            "type": "string",
            "description": "Profile picture URL"
          },
          "status": {
            "type": "string",
            "enum": [
              "online",
              "offline",
              "away",
              "busy"
            ],
            "description": "User status"
          },
          "botEnabled": {
            "type": "boolean",
            "description": "Whether AI bot auto-reply is enabled"
          },
          "botPersonality": {
            "type": "string",
            "enum": [
              "friendly",
              "professional",
              "sarcastic",
              "creative",
              "analytical",
              "mentor",
              "companion"
            ],
            "description": "Preferred AI bot personality"
          },
          "isAdmin": {
            "type": "boolean",
            "description": "Whether user has admin privileges"
          },
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User's full name"
          },
          "bio": {
            "type": "string",
            "description": "User bio"
          },
          "blueTick": {
            "type": "boolean",
            "description": "Blue tick verification status"
          },
          "goldenTick": {
            "type": "boolean",
            "description": "Golden tick verification status"
          },
          "emailVerification": {
            "type": "object",
            "properties": {
              "isVerified": {
                "type": "boolean"
              }
            }
          },
          "profileCompletion": {
            "type": "object",
            "properties": {
              "completionPercentage": {
                "type": "number"
              },
              "hasProfilePicture": {
                "type": "boolean"
              },
              "hasBio": {
                "type": "boolean"
              },
              "hasFollowers": {
                "type": "boolean"
              },
              "hasVerifiedEmail": {
                "type": "boolean"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": "507f1f77bcf86cd799439011",
          "name": "John Doe",
          "username": "johndoe",
          "email": "john@example.com",
          "profilePicture": "https://example.com/profile.jpg",
          "bio": "Tech enthusiast",
          "blueTick": true,
          "goldenTick": false,
          "emailVerification": {
            "isVerified": true
          },
          "profileCompletion": {
            "completionPercentage": 100,
            "hasProfilePicture": true,
            "hasBio": true,
            "hasFollowers": true,
            "hasWalletConnected": true,
            "hasVerifiedEmail": true
          },
          "createdAt": "2024-01-01T12:00:00Z",
          "updatedAt": "2024-01-01T12:00:00Z"
        }
      },
...
      "Community": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "The community ID"
          },
          "name": {
            "type": "string",
            "description": "The community name"
          },
          "description": {
            "type": "string",
            "description": "Community description"
          },
          "profile": {
            "type": "object",
            "properties": {
              "avatar": {
                "type": "string",
                "description": "URL to community avatar"
              },
              "coverPhoto": {
                "type": "string",
                "description": "URL to community cover photo"
              },
              "website": {
                "type": "string",
                "description": "Community website URL"
              },
              "location": {
                "type": "string",
                "description": "Community location"
              }
            }
          },
          "owner": {
            "type": "string",
            "description": "ID of the community owner"
          },
          "moderators": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of moderator user IDs"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "string",
                  "description": "User ID"
                },
                "joinedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "member",
                    "moderator"
                  ]
                }
              }
            }
          },
          "settings": {
            "type": "object",
            "properties": {
              "isPrivate": {
                "type": "boolean",
                "default": false
              },
              "requireApproval": {
                "type": "boolean",
                "default": false
              },
              "canPost": {
                "type": "boolean",
                "default": true
              },
              "canComment": {
                "type": "boolean",
                "default": true
              },
              "payment": {
                "type": "object",
                "properties": {
                  "isPaidCommunity": {
                    "type": "boolean",
                    "default": false
                  },
                  "price": {
                    "type": "number",
                    "default": 0
                  },
                  "currency": {
                    "type": "string",
                    "default": "USD"
                  },
                  "subscriptionType": {
                    "type": "string",
                    "enum": [
                      "one_time",
                      "monthly",
                      "yearly"
                    ],
                    "default": "one_time"
                  },
                  "paymentAddress": {
                    "type": "string"
                  }
                }
              },
              "postSettings": {
                "type": "object",
                "properties": {
                  "allowImages": {
                    "type": "boolean",
                    "default": true
                  },
                  "allowVideos": {
                    "type": "boolean",
                    "default": true
                  },
                  "allowLinks": {
                    "type": "boolean",
                    "default": true
                  },
                  "requireApproval": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          },
          "memberCount": {
            "type": "number",
            "description": "Number of community members"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CommunityPost": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "The post ID"
          },
          "content": {
            "type": "string",
            "description": "Post content"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "community": {
            "$ref": "#/components/schemas/Community"
          },
          "media": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "image",
                    "video",
                    "file"
                  ]
                }
              }
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of user IDs who liked the post"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of comment IDs"
          },
          "shares": {
            "type": "number",
            "default": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "archived"
            ],
            "default": "active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/communities/{id}/posts": {
      "post": {
        "summary": "Create a post in community",
        "description": "Create a post in a community. Requires active membership. For paid communities, requires active subscription. Requires authentication and membership authorization.",
        "tags": [
          "Community Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Community ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Post content"
                  },
                  "media": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "maxItems": 10,
                    "description": "Media files (images, videos)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityPost"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid community ID"
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "403": {
            "description": "Not authorized to post in this community"
          },
          "404": {
            "description": "Community not found"
          }
        }
      },
      "get": {
        "summary": "Get community posts",
        "description": "Retrieve posts from a community. Private communities require membership. Paid communities require active subscription. Requires authentication.",
        "tags": [
          "Community Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Community ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of community posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CommunityPost"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid community ID format"
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "403": {
            "description": "Not authorized to view posts in this community"
          },
          "404": {
            "description": "Community not found"
          }
        }
      }
    },
...
  }
}